Add a new AssignedSample type
This seems like a lot of work for no gain right now. But in order to be
able to attach additional data to the collection samples, we need a type
that wraps the Sample type. This will allow us to add per-collection
notes to the sample (think: keeping notes about status of the project,
the results of a planting, germination notes, etc).

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>