web: Add detail page for collection samples
This is different than the sample itself. It is the information about the
sample within the collection. I admit this is a little bit confusing, so
it may change in the future. In the meantime, this is where you can find
collection-specific notes, etc. The UI is very rough right now.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>