web: Add a note badge to allocations in collection details
We query the latest note for an allocation when we query the allocations
for the collection. So add a little badge to the collection list so that
we can see the category of the latest note for that allocated sample.
This helps keep track of status a bit when planting, etc.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>