lib: Make source description default when querying
Previously when we didn't use the sqlx(rename) on the field, I believe
the field would just default to None when we didn't specifically query
the column. But now it doesn't seem to do that, which was breaking the
"query all child samples" feature on the taxonomy pages. Adding the
'default' attribute makes it work again.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>