Refactor the loadable db object stuff significantly
Instead of making Loadable have Default trait bounds, and implement
new_loadable() to create an default 'hollow' object with only the id
set, move the 'hollow' implementation up a level. Introduce ExternalRef,
which is an enum that can be either a single ID field, or it can be an
full Loadable object loaded from the database.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>