web: add filtering/sorting of samples from project view
Add a sort form to control how the sample list is sorted. This also led
to a bit of a refactor related to how the filtering happens and led to a
discovery that I don't need to use a different route for the filter
request. I can just detect the HX-Request header to determine whether
the request was a full web request or triggered by the form and only
return the fragment of the page in the latter case.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>