web: start handling htmx error responses
Use the 'response-targets' htmx extension which allows you to specify a
different target to swap out error responses int.

Now we can return error status from htmx requests and a bit of error
html to insert into the page rather than re-rendering the entire for
(for example) with an error message embedded.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>