libseed: Loadable: remove fetch_* APIs
Many of the types that implement the Loadable trait already had a
fetch() method (and often several additional fetch_*() methods). Often,
the Loadable::load() method was actually implemented by calling fetch().
Since fetch() and load() did the same thing, this API duplication was
kind of confusing. So just standardize on load()/load_*() for these APIs
and remove the fetch() APIs.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>