db: Add a new history table for samples
I used sqlite_history from Simon Willison (https://github.com/simonw/
sqlite-history) to generate a history table and associated triggers to
track the history of samples in the database. This commit doesn't make
any use of the change history, it only stores it in the database. If we
want to display it in the website later, we'll need to add that support
later.

Resolves: https://github.com/jonner/seedcollection/issues/46

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>