seedctl: add `admin database init` command
This command will initialize a new database starting from an ITIS
taxonomy database. It can download it directly or use a local database
that is already downloaded. It will perform all migrations and will also
add an administrator user to the database.

The Makefile target `prepare-db` will switch to using this new command
instead of doing it within the makefile.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>