lib: optimize db on init and upgrade
The database has a lot of unnecessary data and we should clean this data
after initializing the database and after upgrading to a new database so
that it doesn't use up too much disk space. Also make sure that the
phylo_sort_seq field gets updated on both init and upgrade.

Fixes #67
Fixes #59

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>