lib: Fix Source::FromRow
Previously we were falling back to using the 'tsn' field for the
ExternalRef::Stub object. But this must have been a copy/paste error
since it should be `srcid`. Unfortunately the database currently uses
a TEXT type for srcid in the sc_samples table, so it requires some
additional parsing and fallback behavior.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>