lib: change Loadable::set_id() to set_invalid()
We use the id of the object to determine whether the object is valid
(i.e. whether it is currently stored in the database), so don't let
users set arbitrary Ids on the object. The Id should only be set when
loading from the DB or inserting into the db to ensure that things are
consistent.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>