web: simplify error handling again
This actually reverts a bunch of stuff I previously did because I found
a different way that makes more sense and is a lot more concise. Now,
rely on the error mapping to return the flash messages and just return
Error results from the web handlers as much as possible.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>