web: fix add-to-project form refresh after success
Since we're returning the flash messages out-of-band now, the list
of possible samples to add was not being updated (to remove the
just-added sample) after a successful submission. Fix this by sending
an 'Hx-Trigger' header back in the success response and make the page
re-query the sample list in that case.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>