web: simplify Project::add_sample()
Don't show multiple flash messages when multiple samples fail to be
added to the collection. Also, when sending an error FlashMessage to the
client in response to an htmx request, use HX-Reswap and HX-Retarget to
try to ensure that they get added to the proper div. Also add a new
generic OperationFailed error to handle things like this.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>