web: make project-add-samples sort work reliably
When no 'sort' query field is supplied, we use a default value for the
db query, but we don't tell the template what that default value is,
so that when the user than changes something in the sort/filter menu,
it sends the request with a random sort field which doesn't match the
default used in the rust code. That causes the results to jump around if
you simply tick the 'show empty samples' box. To prevent his, update the
`params` object to contain the default value if none was specified so
that the template knows how to set up its form correctly.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>