db: convert db/germination/gen-sql.py to rust
This script reads a csv file with a list of species with germination
codes and updates the germination codes in the database. I refactored
the 'native status' code to make it more general so that much of it can
be re-used for the germination code stuff as well. The command names
changed slightly.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>