web: Allow app prefix (e.g. /app/...) to be configured
Previously it was set as a compile time constant APP_PREFIX. But this
isn't generally useful. It should either be removed or configurable. We
made it configurable via the config.yaml `public_address` field, (which
used to be called `public_base_url`), which can now have a trailing
path element.

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>