web: Properly verify user email
When a user visits the verify url that they received in their email and
then submits the 'verify' form, we were making the user-verification
request as confirmed, but we were not properly updating the user status
so their profile still indicated that the email had not been verified.
Add a test and also fix the bug.

Fixes: https://github.com/jonner/seedcollection/issues/142

Signed-off-by: Jonathon Jongsma <jonathon@quotidian.org>